concept-runner-mcp/pyproject.toml
[project]
name = "concept-runner-mcp"
version = "0.1.0"
description = "Concept Runner MCP Server - turn research ideas into markdown articles with PubMed papers"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "sqlmodel>=0.0.22",
    "google-genai>=1.0.0",
    "biopython>=1.84",
    "httpx>=0.27.0",
    "tavily-python>=0.5.0",
]

[project.scripts]
concept-runner-mcp = "concept_runner_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/concept_runner_mcp"]
